Moved some pieces around
